browsable (#15482)
CAMEL-21183: BrowsableEndpoint should have limit/filter to optimize returned data from component implementations.

+++ core/camel-management/src/test/java/org/apache/camel/management/ManagedLoadBalancerTest.java
@@ -45,11 +45,11 @@ public class ManagedLoadBalancerTest extends ManagementTestSupport {
         MBeanServer mbeanServer = getMBeanServer();
 
         ObjectName name = getCamelObjectName(TYPE_ENDPOINT, "mock://a");
-        Long queueSize = (Long) mbeanServer.invoke(name, "queueSize", null, null);
+        Integer queueSize = (Integer) mbeanServer.invoke(name, "queueSize", null, null);
         assertEquals(2, queueSize.intValue());
 
         name = getCamelObjectName(TYPE_ENDPOINT, "mock://b");
-        queueSize = (Long) mbeanServer.invoke(name, "queueSize", null, null);
+        queueSize = (Integer) mbeanServer.invoke(name, "queueSize", null, null);
         assertEquals(1, queueSize.intValue());
 
         name = getCamelObjectName(TYPE_PROCESSOR, "myBalancer");
